feat(channel): create basic channel-based broadcast
This adds a basic mechanism to broadcast messages to connections
listening on specific channels. This is achieved through the
`ChannelMap` struct which, as the name suggests, contains a mapping
between channel names and a list of connections that are listening on
the given channel.

Currently, there is no standardized way for a connection to 'join' or
'leave' a channel. In general, this still lacks much of the basic
functionality, but is rather simply to put the `ChannelMap` struct in
place.