perf: remove unnecessary Arc + Mutex around ChannelMap
Now that the ChannelMap lives by itself in its own task, we don't need
any Mutex or Arc around the channels as it will only ever be access by a
single thread. This will definitely need to be added back at some point,
but the decoupled architecture should be much easier to grow around.