feat: rudimentary performance testing and metrics collection
Adds a rudimentary way to keep track of numeric metrics, in this case
just the number of current connections. Also adds a way of doing a
stress-test to see the maximum number of open connections the server can
handle using a swarm of docker containers.

Currently, we can handle ~70k open connections before OOM-ing on 512M of
memory which isn't half bad!