feat: return result from each message opreation
We now send back the result of whatever operation a certain message
performed. For example, we would now return an error back from the
send_command helper if a Broadcast message failed. This will eventually
be used to enable HTTP API requests to be passed as messages to the
ChannelRouter and still be able to retrieve the result.